pyproject.toml
[project]
name = "copy-that"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Adam Brent", email = "adam.brent@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.5",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
    "typer>=0.24.1",
]

[project.scripts]
copy-that = "copy_that.main:main"

[build-system]
requires = ["uv_build>=0.9.30,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
